+++ coref/CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## Version 1.2
+* layout improved
+* keep sentence in chronological order (in chainlists)
+* sentences numbers corrected if "showfrom" is used
+* tests updated
+* help text updated
+
 ## Version 1.1
 * also show bridging if present in data (whole/part and superset/member relations)